Guessing game
A number guessing game using the random library in Python.The users is entering a number until they got it right.With each entry a guess counter is incrementing and at the end there will be a statement with how many guesses you had until you got the correct number. Also after each guess you are told if you were above or below the correct number.